Moved is_self_inverse to unitary
